+++ samplers/samplers.go
@@ -372,7 +372,7 @@ type Set struct {
 
 // Sample checks if the supplied value has is already in the filter. If not, it increments
 // the counter!
-func (s *Set) Sample(sample string, sampleRate float32) {
+func (s *Set) Sample(sample string) {
 	s.Hll.Insert([]byte(sample))
 }